small tidy up and enforcing uint64

+++ events/2024/quests/quest11/quest11_test.go
@@ -13,7 +13,7 @@ func Test1(t *testing.T) {
 
 	rules := parse_data(data)
 	got := days(rules, "A", 4)
-	var want uint = 8
+	var want uint64 = 8
 
 	if got != want {
 		t.Errorf("got %d, wanted %d", got, want)
@@ -29,7 +29,7 @@ func Test2(t *testing.T) {
 
 	rules := parse_data(data)
 	got := all_starts(rules)
-	var want uint = 268815
+	var want uint64 = 268815
 
 	if got != want {
 		t.Errorf("got %d, wanted %d", got, want)
